refactor: Remove transpilation with Peasant
