+++ packages/fury-adapter-apib-serializer/.travis.yml
@@ -6,7 +6,7 @@ node_js:
 - '8'
 - '10'
 script:
-- npm run ci
+- npm test
 after_script:
 - npm install -g codeclimate-test-reporter
 - npm run cover
